+++ demo/pages/hierarchy-template/index.js
@@ -1,10 +1,30 @@
-Controller.$inject = ['$http'];
-export default function Controller($http) {
-	const ctrl = this;
-	this.rows = [];
+Controller.$inject = ['$http', 'qgrid'];
+export default function Controller($http, qgrid) {
+	const model = qgrid.model();
+	const service = qgrid.service(model);
+
+	this.gridModel = model;
 
 	$http.get('data/people/100.json')
-		.then(function (response) {
-			ctrl.rows = response.data;
-		});
+		.then(response =>
+			model.data({
+				rows: response.data
+			})
+		);
+
+	let shouldExpand = true;
+	const toggleAll = nodes => nodes.forEach(node => {
+		if (node.type === 'group') {
+			node.state.expand = shouldExpand;
+			toggleAll(node.children);
+		}
+	});
+
+	this.toggleCollapse = () => {
+		const nodes = model.view().nodes;
+		toggleAll(nodes);
+
+		shouldExpand = !shouldExpand;
+		service.invalidate('app', {}, qgrid.pipeUnit.group);
+	};
 }